Move AddActionItem component to RetroItemWithAddAction folder & use Textarea in AddActionItem
