+++ ui/src/App/Team/Retro/ThoughtColumn/RetroItemWithAddAction/RetroItemWithAddAction.scss
@@ -18,10 +18,6 @@
 @use '../../../../../Styles/main';
 
 .retro-item-with-add-action {
-	.retro-item {
-		margin-bottom: 30px;
-	}
-
 	&.creating-action {
 		.retro-item {
 			box-shadow: 0 19px 38px rgba(main.$black, 0.3) !important;
@@ -39,7 +35,7 @@
 	.add-action-item-button {
 		display: block;
 		height: 35px;
-		margin: 0 auto;
+		margin: 1.8rem auto 0;
 		padding: 8px 16px;
 
 		font-weight: bolder;
